back

+++ src/main/java/com/example/proyecto/curso/domail/CursoService.java
@@ -1,5 +1,6 @@
 package com.example.proyecto.curso.domail;
 
+import com.example.proyecto.auth.config.AuthorizationConfig;
 import com.example.proyecto.carrera.domail.Carrera;
 import com.example.proyecto.carrera.infrastructure.CarreraRepository;
 import com.example.proyecto.curso.dto.CursoRequestDto;
@@ -7,13 +8,19 @@ import com.example.proyecto.curso.dto.CursoResponseDto;
 import com.example.proyecto.curso.infrastructure.CursoRepository;
 import com.example.proyecto.exception.ResourceConflictException;
 import com.example.proyecto.exception.ResourceNotFoundException;
+import com.example.proyecto.exception.UnauthorizeOperationException;
+import com.example.proyecto.post.domail.Post;
 import com.example.proyecto.usuario.domail.Usuario;
 import com.example.proyecto.usuario.dto.UsuarioResponseDto;
+import com.example.proyecto.usuario.infrastructure.UsuarioRepository;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CursoService {
@@ -21,40 +28,64 @@ public class CursoService {
     private final CursoRepository cursoRepository;
     private final CarreraRepository carreraRepository;
     private final ModelMapper modelMapper;
+    private final AuthorizationConfig authorizationConfig;
+    private final UsuarioRepository usuarioRepository;
     public CursoService(CursoRepository cursoRepository,
                         CarreraRepository carreraRepository,
-                        ModelMapper modelMapper) {
+                        ModelMapper modelMapper,
+                        AuthorizationConfig authorizationConfig,
+                        UsuarioRepository usuarioRepository) {
         this.cursoRepository = cursoRepository;
         this.carreraRepository = carreraRepository;
         this.modelMapper = modelMapper;
+        this.authorizationConfig = authorizationConfig;
+        this.usuarioRepository = usuarioRepository;
     }
 
 
     public CursoResponseDto createCurso(Long carreraId,CursoRequestDto requestDto) {
 
+        String userEmail = authorizationConfig.getCurrentUserEmail();
+        if(userEmail == null){
+            throw new ResourceConflictException("No se encuentra AUTORIZADO!!");
+        }
+        Usuario usuario = usuarioRepository.findByEmail(userEmail)
+                .orElseThrow(() -> new ResourceNotFoundException("Usuario no encontrado"));
+
+        Carrera carrera = carreraRepository.findById(carreraId)
+                .orElseThrow(() -> new ResourceNotFoundException("Carrera no encontrada"));
+
+        if (!usuario.getCarreras().contains(carrera)) {
+            throw new UnauthorizeOperationException("No puedes actualizar. No está inscrito en la carrera.");
+        }
+
         if (cursoRepository.findByNombre(requestDto.getNombre()).isPresent()) {
             throw new ResourceConflictException("Curso ya existente");
         }
-        Carrera carrera = carreraRepository.findById(carreraId)
-                .orElseThrow(() -> new ResourceNotFoundException("Carrera no encontrada"));
+
 
         Curso curso = new Curso();
         modelMapper.map(requestDto, curso);
 
         curso.setCarrera(carrera);
-        //carrera.getCursos().add(curso);
 
         cursoRepository.save(curso);
         return modelMapper.map(curso, CursoResponseDto.class);
     }
 
-    public List<CursoResponseDto> retornarByCarrera(){
-        List<Curso> cursos = cursoRepository.findAll();
-        List<CursoResponseDto> cursosResponseDto = new ArrayList<>();
-        for (Curso curso : cursos) {
-            cursosResponseDto.add(modelMapper.map(curso, CursoResponseDto.class));
+    public List<CursoResponseDto> obtenerCursosPorCarrera(Long carreraId) {
+        if (!carreraRepository.existsById(carreraId)) {
+            throw new ResourceNotFoundException("Carrera no encontrada con el ID: " + carreraId);
         }
-        return cursosResponseDto;
+
+        List<Curso> curso = cursoRepository.findByCarreraId(carreraId);
+        List<CursoResponseDto> cursodto = new ArrayList<>();
+
+        for (Curso c : curso) {
+            cursodto.add(modelMapper.map(c, CursoResponseDto.class));
+        }
+
+        return cursodto;
     }
 
     public CursoResponseDto getCursoById(Long id) {
@@ -63,22 +94,49 @@ public class CursoService {
         return modelMapper.map(curso, CursoResponseDto.class);
     }
 
-    public CursoResponseDto updateCurso(Long carreraId,Long id, CursoRequestDto requestDto) {
+    public CursoResponseDto updateCurso(Long id, CursoRequestDto requestDto) {
+
+        String userEmail = authorizationConfig.getCurrentUserEmail();
+
+        if(userEmail == null){
+            throw new UnauthorizeOperationException("No se encuentra AUTORIZADO!!");
+        }
+
+        Usuario usuario = usuarioRepository.findByEmail(userEmail)
+                .orElseThrow(() -> new ResourceNotFoundException("Usuario no encontrado"));
+
         Curso curso = cursoRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Curso no encontrado"));
-        modelMapper.map(requestDto, curso);
 
-        Carrera carrera = carreraRepository.findById(carreraId)
-                .orElseThrow(() -> new ResourceNotFoundException("Carrera no encontrada"));
-        curso.setCarrera(carrera);
+        if (!usuario.getCarreras().contains(curso.getCarrera())) {
+            throw new UnauthorizeOperationException("No puedes actualizar. No está inscrito en la carrera.");
+        }
 
+        curso.setNombre(requestDto.getNombre());
         cursoRepository.save(curso);
+
+
         return modelMapper.map(curso, CursoResponseDto.class);
     }
 
     public void deleteCurso(Long id) {
+
+        String userEmail = authorizationConfig.getCurrentUserEmail();
+
+        if(userEmail == null){
+            throw new UnauthorizeOperationException("No se encuentra AUTORIZADO!!");
+        }
+
+        Usuario usuario = usuarioRepository.findByEmail(userEmail)
+                .orElseThrow(() -> new ResourceNotFoundException("Usuario no encontrado"));
+
         Curso curso = cursoRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Curso no encontrado"));
+
+        if (!usuario.getCarreras().contains(curso.getCarrera())) {
+            throw new UnauthorizeOperationException("No puedes Eliminar. No está inscrito en la carrera.");
+        }
+
         cursoRepository.delete(curso);
     }
 }
